+++ internal/backend/util/foreground_unix.go
@@ -1,7 +1,7 @@
 //go:build !aix && !solaris && !windows
 // +build !aix,!solaris,!windows
 
-package backend
+package util
 
 import (
 	"os"
